Grille HEALPix générique

+++ src/healpix/essentials/HealpixUtils.java
@@ -44,7 +44,7 @@ public class HealpixUtils
       @return the integer {@code n} which fulfills {@code n^2<=arg<(n+1)^2} */
   static public int isqrt(long arg)
     {
-    long res = (long)Math.sqrt(((double)arg)+0.5);
+    long res = (long)Math.sqrt((arg)+0.5);
     if (arg<(1L<<50)) return (int)res;
     if (res*res>arg)
       --res;
